Refactor: Removed all the public functions to the source file from the header the file
